refactor(UrlField): merge from VideoField
